Optimiere Laufzeitperformance
- Alle Operationen nutzen float statt double
  (EV3 hat keine FP-Unit!)
- Übersetzungsverhältnisse, Faktoren für mm<->Grad
  Umrechnungen bei Instantiierung berechnen statt
  bei jedem Aufruf
- Einzelne Berechnungen umstellen

Prüfen, ob Umstellung auf Integeroperationen
möglich bzw notwendig ist

+++ src/plott3r_V1_solved/IUebersetzung.java
@@ -2,7 +2,7 @@ package plott3r_V1_solved;
 
 public interface IUebersetzung {
 
-	public double getUebersetzungsverhaeltnis();
+	public float getUebersetzungsverhaeltnis();
 
 	public boolean isAntriebsUmkehrung();
