+++ src/positions/Position2D.java
@@ -4,52 +4,53 @@ import util.MathHelper;
 
 public class Position2D {
 
-	private double x, y;
+	private float x, y;
 
 	// NEU: Add empty constructor
 	public Position2D() {
-		this.x = Double.NaN;
-		this.y = Double.NaN;
+		this.x = Float.NaN;
+		this.y = Float.NaN;
 	}
-	
-	public Position2D(double x, double y) {
+
+	public Position2D(float x, float y) {
 		this.x = x;
 		this.y = y;
 	}
 
 	@Override
 	public boolean equals(Object obj) {
-		if (this.getXRounded() == ((Position2D) obj).getXRounded() && this.getYRounded() == ((Position2D) obj).getYRounded())
+		if (this.getXRounded() == ((Position2D) obj).getXRounded()
+				&& this.getYRounded() == ((Position2D) obj).getYRounded())
 			return true;
 
 		return false;
 	}
 
-	public double getX() {
+	public float getX() {
 		return x;
 	}
 
-	private double getXRounded() {
+	private float getXRounded() {
 		return round(this.getX());
 	}
 
-	public double getY() {
+	public float getY() {
 		return y;
 	}
 
-	private double getYRounded() {
+	private float getYRounded() {
 		return round(this.getY());
 	}
 
-	private double round(double d) {
-		return MathHelper.round(d, 2);
+	private float round(float d) {
+		return (float) MathHelper.round(d, 2);
 	}
 
-	public void setX(double x) {
+	public void setX(float x) {
 		this.x = x;
 	}
 
-	public void setY(double y) {
+	public void setY(float y) {
 		this.y = y;
 	}
